+++ qiskit_ibm_runtime/utils/estimator_result_decoder.py
@@ -15,8 +15,8 @@
 from typing import Dict
 import numpy as np
 
-from .program.result_decoder import ResultDecoder
-from .qiskit.primitives import EstimatorResult
+from ..program.result_decoder import ResultDecoder
+from ..qiskit.primitives import EstimatorResult
 
 
 class EstimatorResultDecoder(ResultDecoder):
